Share project "car-pool" into "https://car-pool.googlecode.com/svn"
Made a basic start with a loggin html page and a Authenticate Servlet to display the form inputs from loggin.html.

Found a servlet tester in the jetty project and am going to use this as a basis for testing my servlets. See http://roberthanson.blogspot.com/2007/12/testing-servlets-with-junit.html for an example of using this library.