Separated OpenId registration from Username and password registration.  It's now simpler and therefore less prone to bugs.
Also modified logout.jsp to get rid of the "signedin" attribute.  It may not be neccessary but I did it just for completeness sake. 

Also added a link to the register page in index.jsp, for Username and password registration.  OpenId registration does not need this link as it will go to the registration page if the "openid_url" is not in the database.