A new JDepend report.
OpenIdConsumer now uses a ProxySelector I wrote myself setting to the default ProxySelectory.  This is the class all network sockets call on when they try and establish a connection, and see what proxies, if any, they need to use.

Wrote ProxyConfig to plug into the spot where the default ProxySelector is.  At the moment I have hardcoded the default massey proxies into this class so we might want to deploy it soon and try it out with openid login.

ProxyData is stored in ProxyConfig to hold data on the proxy as well as the Proxy class that stores info about the proxy.  This includes things like failedCount, which is how many attempts have been made and failed to connect to the proxy.

There are a few other minor changes in other files.