Registration form validation now works.
Made a change in CheckUserNameAvailable so that instead of println it uses print, which allows me to check the returned text from that servlet more easily.

The forms are checked to see if userName and email input fields have values and if you are using username and password registration checks to see if there is values entered in both password fields and also if they match or not.