Added openid image to openid log in form.
Created a BlurredImage servlet along with RandomTextGenerator to allow the production of images with text in them for verification and safeguarding the registration of users.

Modified the registration processing servlets to check this new input.

Modified the javascript form form input validation to check to see if input has been put into this new input field.

Added a new field into the registration forms for verification text to be entered into and added the image to appear above it.