Not sure what I did to some of these files, but they would be minor changes.
Added a Email and SMTP class.  You create a instance of Email, fill it with the right details and then use the static method SMTP.send(email).  At the moment the SMTP.send is a skeleton, but you should be able to write code to use this feature and I will get it to work in the future.

I also removed a filter that is redundant in the web.xml file.