Suppressed some warnings.
Yet again fixed the character case issue where one should be using Matches not matches.

Had to also fix a error which has occured before where the ResultSet rs is null and the code tries to do something like rs.next() and because it's null it throws a null pointer exception.