Implemented emailing from various parts of the code.
Created a RideUserStack, to provide a easy way of getting a list of users in a ride other than the one offering the ride.  I don't know how else to do that one.

Have stated implementing the registration of a admin user, so anyone trying to log in as admin with the '12weak34' password will be forwarded to adminregistration.jsp, which is a work in progress.